Make <1.21.2 container id types consistent with Vanilla again ... that's to say, to make them inconsistent thanks to special ids: Byte: - CONTAINER_SET_SLOT - CONTAINER_CLICK - PLACE_GHOST_RECIPE - PLACE_RECIPE - CONTAINER_BUTTON_CLICK - CONTAINER_CLOSE (serverbound) - CONTAINER_ACK (serverbound) Unsigned byte (unchanged): - OPEN_SCREEN - CONTAINER_SET_CONTENT - CONTAINER_SET_DATA - HORSE_SCREEN_OPEN - CONTAINER_CLOSE (clientbound) - CONTAINER_ACK (clientbound) Short (unchanged): - SET_CREATIVE_MODE_SLOT
This doesn't change much in practice, but means it matches the Vanilla reading

+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_8to1_9/storage/EntityTracker1_9.java
@@ -128,7 +128,7 @@ public class EntityTracker1_9 extends EntityTrackerBase {
 
         // Get item in new selected slot
         int inventorySlot = this.heldItemSlot + 36; // Hotbar slot index to inventory slot
-        int itemIdentifier = inventoryTracker.getItemId((short) 0, (short) inventorySlot);
+        int itemIdentifier = inventoryTracker.getItemId(0, (short) inventorySlot);
 
         return Protocol1_8To1_9.isSword(itemIdentifier);
     }
